+++ app/components/EnhancedTableToolbar.js
@@ -84,11 +84,11 @@ class EnhancedTableToolbar extends Component<Props> {
       >
         <div className={classes.title}>
           {numSelected > 0 ? (
-            <Typography color="inherit" variant="subheading">
+            <Typography color="inherit" variant="subtitle1">
               {numSelected} selected
             </Typography>
           ) : (
-            <Typography variant="title" id="tableTitle" />
+            <Typography variant="h6" id="tableTitle" />
           )}
         </div>
         <div className={classes.spacer} />