+++ app/components/ManageBundleManifestResourcesDialog.js
@@ -36,7 +36,6 @@ import { closeResourceManager,
 import { selectItemsToPaste, pasteItems, clearClipboard } from '../actions/clipboard.actions';
 import { downloadResources, removeResources, getEntryRevisions, createBundleFromDBL, selectBundleEntryRevision } from '../actions/bundle.actions';
 import { openMetadataFile } from '../actions/bundleEditMetadata.actions';
-import rowStyles from './DBLEntryRow.css';
 import EnhancedTable from './EnhancedTable';
 import EnhancedTableToolbar from './EnhancedTableToolbar';
 import { utilities } from '../utils/utilities';
@@ -46,6 +45,7 @@ import ConfirmButton from '../components/ConfirmButton';
 import CopyForPasteButton from './CopyForPasteButton';
 import PasteButton from './PasteButton';
 import MapperTable from '../components/MapperTable';
+import EntryTitle from '../components/EntryTitle';
 
 const { dialog } = require('electron').remote;
 const { shell } = require('electron');
@@ -1311,10 +1311,10 @@ class ManageBundleManifestResourcesDialog extends Component<Props> {
     return Object.entries(wizardsResults).map(([wizardName, results]) =>
       (
         <React.Fragment key="frag">
-          <Typography key={`${wizardName}-description`} variant="subheading" color="inherit" paragraph>
+          <Typography key={`${wizardName}-description`} variant="subtitle1" color="inherit" paragraph>
             • <b>{results.description}</b> ({wizardName}):
           </Typography>
-          <Typography key={`${wizardName}-documentation`} variant="subheading" color="inherit" style={{ marginLeft: '20px' }} paragraph>
+          <Typography key={`${wizardName}-documentation`} variant="subtitle1" color="inherit" style={{ marginLeft: '20px' }} paragraph>
             {results.documentation}
           </Typography>
         </React.Fragment>
@@ -1475,10 +1475,9 @@ class ManageBundleManifestResourcesDialog extends Component<Props> {
     } = this.props;
     const { storedResources } = this.getSelectedResourcesByStatus();
     const { displayAs = {} } = origBundle;
-    const { languageAndCountry, name, revision } = displayAs;
+    const { revision } = displayAs;
     const modeUi = this.modeUi();
     const isModifyFilesMode = this.isModifyFilesMode();
-    const mediumIconMarginRight = ux.getMediumIcon(origBundle.medium);
     const { status, parent, dblId } = origBundle;
     const revBackground =
       ux.getDblRowBackgroundColor(false, classes, status, revision, parent, dblId);
@@ -1491,8 +1490,8 @@ class ManageBundleManifestResourcesDialog extends Component<Props> {
                 <CloseIcon />
               </IconButton>
               <FolderOpen color="inherit" className={classNames(classes.leftIcon)} />
-              <Typography variant="title" color="inherit">
-                {modeUi.appBar.title}: {mediumIconMarginRight} <span className={rowStyles.languageAndCountryLabel}>{languageAndCountry} </span> {name}
+              <Typography variant="h6" color="inherit">
+                {modeUi.appBar.title}: {<EntryTitle bundle={origBundle} />}
               </Typography>
               <Tooltip title={this.props.entryPageUrl}>
                 <Button onClick={this.onOpenDBLEntryLink} className={classNames(classes.button, revBackground)}>
@@ -1519,7 +1518,7 @@ class ManageBundleManifestResourcesDialog extends Component<Props> {
           </AppBar>
           {isModifyFilesMode && publicationsHealthMessage &&
             <Toolbar className={classes.errorBar}>
-              <Typography variant="subheading" color="inherit">
+              <Typography variant="subtitle1" color="inherit">
                 {publicationsHealthMessage}
               </Typography>
               <div style={{ paddingLeft: '10px' }} />
@@ -1535,7 +1534,7 @@ class ManageBundleManifestResourcesDialog extends Component<Props> {
           {!loading && isModifyFilesMode && publicationsHealthSuccessMessage &&
             <Card className={classes.successBar} raised>
               <CardContent>
-                <Typography key="pubhealthSuccessMessage" variant="subheading" color="inherit" gutterBottom>
+                <Typography key="pubhealthSuccessMessage" variant="subtitle1" color="inherit" gutterBottom>
                   {publicationsHealthSuccessMessage}
                 </Typography>
                 {this.renderWizardsResults()}