+++ app/components/PasteButton.js
@@ -1,30 +1,35 @@
-import React, { Component } from 'react';
+import React, { PureComponent } from 'react';
+import Tooltip from '@material-ui/core/Tooltip';
 import classNames from 'classnames';
 import AssignmentTurnedInIcon from '@material-ui/icons/AssignmentTurnedIn';
 import ConfirmButton from '../components/ConfirmButton';
+import { ux } from '../utils/ux';
 
 type Props = {
   onClick: () => {},
   classes: {},
-  itemsToPaste: [],
-  itemsType: string
+  selectedItemsToPaste: {}
 };
 
-export default class PasteButton extends Component<Props> {
+export default class PasteButton extends PureComponent<Props> {
   props: Props;
 
   render() {
     const {
-      classes, itemsToPaste, itemsType, ...restProps
+      classes, selectedItemsToPaste, ...restProps
     } = this.props;
+    const { items: itemsToPaste, itemsType } = selectedItemsToPaste || {};
+    const clipboardTooltip = ux.getClipboardTooltip(selectedItemsToPaste);
     return (
-      <ConfirmButton
-        classes={classes}
-        {...restProps}
-      >
-        <AssignmentTurnedInIcon className={classNames(classes.leftIcon)} />
-        Paste {itemsToPaste.length ? `(${itemsToPaste.length}) ` : ''} {itemsType}
-      </ConfirmButton>
+      <Tooltip title={clipboardTooltip}>
+        <ConfirmButton
+          classes={classes}
+          {...restProps}
+        >
+          <AssignmentTurnedInIcon className={classNames(classes.leftIcon)} />
+          Paste {itemsToPaste.length ? `(${itemsToPaste.length}) ` : ''} {itemsType}
+        </ConfirmButton>
+      </Tooltip>
     );
   }
 }