+++ app/utils/ux.js
@@ -1,21 +1,16 @@
 import React from 'react';
-import Book from '@material-ui/icons/Book';
-import Headset from '@material-ui/icons/Headset';
-import Videocam from '@material-ui/icons/Videocam';
-import Print from '@material-ui/icons/Print';
-import Grain from '@material-ui/icons/Grain';
 import Badge from '@material-ui/core/Badge';
 import { lighten } from '@material-ui/core/styles/colorManipulator';
 import { bundleService } from '../services/bundle.service';
 
 export const ux = {
-  getMediumIcon,
   getFormattedRevision,
   getDblRowStyles,
   getDblRowBackgroundColor,
   mapColumns,
   getHighlightTheme,
-  conditionallyRenderBadge
+  conditionallyRenderBadge,
+  getClipboardTooltip
 };
 export default ux;
 
@@ -40,15 +35,6 @@ function getFormattedRevision(bundle, insertStr) {
     return `${insertStr}1 (New)`;
   }
   return `${insertStr}${revision}`;
-}
-
-export function getMediumIcon(medium, props = { style: { marginRight: '10px' } }) {
-  return (medium === 'text' && <Book {...props} />)
-  || (medium === 'audio' && <Headset {...props} />)
-  || (medium === 'video' && <Videocam {...props} />)
-  || (medium === 'print' && <Print {...props} />)
-  || (medium === 'braille' && <Grain {...props} />)
-  || (null);
 }
 
 function getDblRowStyles(theme) {
@@ -118,4 +104,12 @@ function conditionallyRenderBadge(props, content, node) {
     return node;
   }
   return <Badge key="badge" {...props} badgeContent={content}>{node}</Badge>;
+}
+
+function getClipboardTooltip(selectedItemsToPaste) {
+  const {
+    bundleId, itemsType, getDisplayAs, getMedium
+  } = selectedItemsToPaste;
+  const clipboardTooltip = bundleId ? `${itemsType} from (${getMedium()}) ${getDisplayAs().name} ${getDisplayAs().revision}` : '';
+  return clipboardTooltip;
 }